revert previous fix for #107
Issue #107 should be rejected as it is much easier to handle
that process from a start script that it is from fetch.

In my experiences testing the implementation of #107 it was
apparent that we could not preset many of the values
(e.g., remote_peer_port ...). Much easier to keep having
`epm fetch` get the blockchain and then do a small
`epm run & sleep 20 && kill $(epm plop pid)` which gives
us much more flexibility when folks just run epm
from containers.